// src/app/models/vehicle.ts
export default interface Vehicle {
    _id: string;
    ownerId: string;
    degem: string;
    manufacturer: string;
    year: number;
    horsepower: number;
    gearbox: string;
    yad: string;
    kilometraj: number;
    location: string;
    description: string;
    price: number;
    color: string;
    moed_aliya_lakvish: number;
    baalut: string;
    ownerName: string;
    ownerPhone: string;
    images: string[];
}